profiles: refactor code structure - using index for components
