+++ client/src/components/index.ts
@@ -1,0 +1,6 @@
+// Re-export all components for convenient importing
+export * from './error';
+export * from './footer';
+export * from './header';
+export * from './layout';
+export * from './user-menu';
