+++ client/src/components/layout/Layout.tsx
@@ -1,6 +1,6 @@
 import { ReactNode, useEffect } from 'react';
-import { Header } from '../header/Header';
-import { Footer } from '../footer/Footer.tsx';
+import { Header } from '../header';
+import { Footer } from '../footer';
 import './Layout.css';
 
 interface LayoutProps {
@@ -10,12 +10,7 @@ interface LayoutProps {
     className?: string;
 }
 
-export const Layout = ({
-    children,
-    headerVariant = 'compact',
-    showFooter = true,
-    className = '',
-}: LayoutProps) => {
+export const Layout = ({ children, headerVariant = 'compact', showFooter = true, className = '' }: LayoutProps) => {
     useEffect(() => {
         if (headerVariant === 'compact') {
             document.body.classList.add('has-fixed-header');
@@ -32,9 +27,7 @@ export const Layout = ({
         <div className={`layout ${className}`}>
             <Header variant={headerVariant} />
 
-            <main
-                className={`layout-main ${headerVariant === 'compact' ? 'main-with-fixed-header' : ''}`}
-            >
+            <main className={`layout-main ${headerVariant === 'compact' ? 'main-with-fixed-header' : ''}`}>
                 {children}
             </main>
